turn intra-doc-link collection into an early pass
